Les tests d'integration marche à merveil et les données sont bien ajoutées dans la base de données

+++ app/src/main/java/com/llab/ligablo/database/dao/etabs/EtablissementDao.java
@@ -1,4 +1,6 @@
 package com.llab.ligablo.database.dao.etabs;
+
+
 import android.arch.lifecycle.LiveData;
 import android.arch.persistence.room.Dao;
 import android.arch.persistence.room.Insert;
@@ -12,10 +14,10 @@ import java.util.List;
 public interface EtablissementDao {
 
     @Query("SELECT * FROM Etablissement WHERE userId = :userId ORDER BY id DESC")
-    LiveData<List<Etablissement>> getEtablissements(int userId);
+    LiveData<List<Etablissement>> getEtablissements(long userId);
 
     @Query("SELECT * FROM Etablissement WHERE userId = :userId AND id=:etabId")
-    LiveData<List<Etablissement>> getEtablissement(int userId,int etabId);
+    LiveData<List<Etablissement>> getEtablissement(long userId, long etabId);
 
     @Insert
     long insertEtablissement(Etablissement etablissement);
@@ -24,5 +26,5 @@ public interface EtablissementDao {
     int updateEtablissement(Etablissement etablissement);
 
     @Query("DELETE FROM Etablissement WHERE id = :etablissementId")
-    int deleteEtablissement(int etablissementId);
+    int deleteEtablissement(long etablissementId);
 }
