+++ app/src/main/java/com/llab/ligablo/database/dao/produits/StockDao.java
@@ -8,7 +8,6 @@ import android.arch.persistence.room.Update;
 
 import com.llab.ligablo.models.produits.Stock;
 
-import java.util.Date;
 import java.util.List;
 
 @Dao
@@ -18,22 +17,22 @@ public interface StockDao {
     LiveData<List<Stock>> getStocks();
 
     @Query("SELECT * FROM Stock WHERE extensionId=:extensionId")
-    LiveData<List<Stock>> getStocks_InExtension(int extensionId);
+    LiveData<List<Stock>> getStocks_InExtension(long extensionId);
 
     @Query("SELECT * FROM Stock WHERE extensionId=:extensionId AND produitId=:productId")
-    LiveData<List<Stock>> getStocks_InExtension_ByProduct(int extensionId,int productId);
+    LiveData<List<Stock>> getStocks_InExtension_ByProduct(long extensionId, long productId);
 
     @Query("SELECT * FROM Stock WHERE extensionId=:extensionId AND  quantite=:QteStock")
-    LiveData<List<Stock>> getStocks_InExtension_ByQuantite(int extensionId,int QteStock);
+    LiveData<List<Stock>> getStocks_InExtension_ByQuantite(long extensionId, long QteStock);
 
     @Query("SELECT * FROM Stock WHERE extensionId=:extensionId AND quantite=:QteStock AND produitId=:productId")
-    LiveData<List<Stock>> getStocks_InExtension_ByQuantiteAndProduct(int extensionId,int QteStock,int productId);
+    LiveData<List<Stock>> getStocks_InExtension_ByQuantiteAndProduct(long extensionId, long QteStock, long productId);
 
     @Query("SELECT * FROM Stock WHERE extensionId=:extensionId AND date=:date")
-    LiveData<List<Stock>> getStocks_InExtensionByDate(int extensionId, String date);
+    LiveData<List<Stock>> getStocks_InExtensionByDate(long extensionId, String date);
 
     @Query("SELECT * FROM Stock WHERE date=:date1 AND extensionId=:extensionId UNION SELECT * FROM Stock WHERE date=:date2 AND extensionId=:extensionId")
-    LiveData<List<Stock>> getStocks_InExtensionBetweenTwoDate(int extensionId,String date1,String date2);
+    LiveData<List<Stock>> getStocks_InExtensionBetweenTwoDate(long extensionId, String date1, String date2);
 
     @Insert
     long insertStock(Stock stock);
@@ -42,5 +41,5 @@ public interface StockDao {
     int updateStock(Stock stock);
 
     @Query("DELETE FROM Stock WHERE id =:stockId")
-    int deleteStock(int stockId);
+    int deleteStock(long stockId);
 }
